comment sync tryouts

+++ src/main/java/br/com/xht/udp/protocol/soapbox/SoapBoxTalk.java
@@ -31,10 +31,25 @@ public class SoapBoxTalk extends UdpTalk {
 		}
 	}
 
+	// private byte pingCalc() {
+	// long diffTime = getDiffTime();
+	// if (diffTime > 1000) {
+	// return (byte) 0xff;
+	// }
+	// float pingCalcFloat = 0.254F * diffTime;
+	// byte pingCalc = new BigDecimal(pingCalcFloat).byteValue();
+	// return pingCalc;
+	// }
+
 	@Override
 	public byte[] getSyncPacket() {
+		// byte[] helloPacket = getHelloPacket();
 		byte[] pingSyncPacket = syncPacket;
-		pingSyncPacket[(pingSyncPacket.length - 6)] = 0x03;
+		// pingSyncPacket[5] = seqSync; // seq from hello part?
+		// pingSyncPacket[6] = pingCalc(); // ping calc?
+		// pingSyncPacket[7] = helloPacket[2]; // hello part, fixed byte?
+		// pingSyncPacket[8] = helloPacket[3]; // hello part, packet loss?
+		pingSyncPacket[(pingSyncPacket.length - 6)] = (byte) 0x03;
 		return pingSyncPacket;
 	}
